Remove unused code after removal of some ruby WS
