+++ server/sonar-server/src/main/java/org/sonar/server/platform/platformlevel/PlatformLevel4.java
@@ -61,7 +61,6 @@ import org.sonar.server.component.ComponentFinder;
 import org.sonar.server.component.ComponentService;
 import org.sonar.server.component.DefaultComponentFinder;
 import org.sonar.server.component.DefaultRubyComponentService;
-import org.sonar.server.component.ws.ComponentJsonWriter;
 import org.sonar.server.component.ws.ComponentsWs;
 import org.sonar.server.component.ws.EventsWs;
 import org.sonar.server.component.ws.ResourcesWs;
@@ -345,16 +344,16 @@ public class PlatformLevel4 extends PlatformLevel {
       IndexDefinitions.class,
       IndexCreator.class,
 
-      // Activity
+    // Activity
       ActivityService.class,
       ActivityIndexDefinition.class,
       ActivityIndexer.class,
       ActivityIndex.class,
 
-      // batch
+    // batch
       BatchWsModule.class,
 
-      // Dashboard
+    // Dashboard
       DashboardsWs.class,
       org.sonar.server.dashboard.ws.ShowAction.class,
       ProjectDefaultDashboard.class,
@@ -392,12 +391,12 @@ public class PlatformLevel4 extends PlatformLevel {
       ProjectIssueFilterWidget.class,
       IssueTagCloudWidget.class,
 
-      // update center
+    // update center
       UpdateCenterClient.class,
       UpdateCenterMatrixFactory.class,
       UpdateCenterWs.class,
 
-      // quality profile
+    // quality profile
       XMLProfileParser.class,
       XMLProfileSerializer.class,
       AnnotationProfileParser.class,
@@ -440,7 +439,7 @@ public class PlatformLevel4 extends PlatformLevel {
       QProfileReset.class,
       RubyQProfileActivityService.class,
 
-      // rule
+    // rule
       AnnotationRuleParser.class,
       XMLRuleParser.class,
       DefaultRuleFinder.class,
@@ -468,17 +467,17 @@ public class PlatformLevel4 extends PlatformLevel {
       RepositoriesAction.class,
       org.sonar.server.rule.ws.AppAction.class,
 
-      // languages
+    // languages
       Languages.class,
       LanguageWs.class,
       org.sonar.server.language.ws.ListAction.class,
 
-      // activity
+    // activity
       ActivitiesWs.class,
       org.sonar.server.activity.ws.SearchAction.class,
       ActivityMapping.class,
 
-      // measure
+    // measure
       MeasureFilterFactory.class,
       MeasureFilterExecutor.class,
       MeasureFilterEngine.class,
@@ -490,14 +489,14 @@ public class PlatformLevel4 extends PlatformLevel {
       DefaultMetricFinder.class,
       TimeMachineWs.class,
 
-      // quality gates
+    // quality gates
       QualityGateDao.class,
       QualityGateConditionDao.class,
       QualityGates.class,
       ProjectQgateAssociationDao.class,
       QgateProjectFinder.class,
 
-      org.sonar.server.qualitygate.ws.ListAction.class,
+    org.sonar.server.qualitygate.ws.ListAction.class,
       org.sonar.server.qualitygate.ws.SearchAction.class,
       org.sonar.server.qualitygate.ws.ShowAction.class,
       org.sonar.server.qualitygate.ws.CreateAction.class,
@@ -514,17 +513,17 @@ public class PlatformLevel4 extends PlatformLevel {
       org.sonar.server.qualitygate.ws.AppAction.class,
       QGatesWs.class,
 
-      // web services
+    // web services
       WebServiceEngine.class,
       ListingWs.class,
 
-      // localization
+    // localization
       L10nWs.class,
 
-      // authentication
+    // authentication
       AuthenticationWs.class,
 
-      // users
+    // users
       SecurityRealmFactory.class,
       DeprecatedUserFinder.class,
       NewUserNotifier.class,
@@ -546,19 +545,19 @@ public class PlatformLevel4 extends PlatformLevel {
       UserIndex.class,
       UserUpdater.class,
 
-      // groups
+    // groups
       GroupMembershipService.class,
       GroupMembershipFinder.class,
       UserGroupsModule.class,
 
-      // permissions
+    // permissions
       PermissionRepository.class,
       PermissionService.class,
       PermissionTemplateService.class,
       PermissionFinder.class,
       PermissionsWsModule.class,
 
-      // components
+    // components
       ProjectsWsModule.class,
       DefaultComponentFinder.class,
       DefaultRubyComponentService.class,
@@ -572,14 +571,13 @@ public class PlatformLevel4 extends PlatformLevel {
       NewAlerts.class,
       NewAlerts.newMetadata(),
       ComponentCleanerService.class,
-      ComponentJsonWriter.class,
 
-      // views
+    // views
       ViewIndexDefinition.class,
       ViewIndexer.class,
       ViewIndex.class,
 
-      // issues
+    // issues
       IssueIndexDefinition.class,
       IssueIndexer.class,
       IssueAuthorizationIndexer.class,
@@ -613,13 +611,13 @@ public class PlatformLevel4 extends PlatformLevel {
       EmailNotificationChannel.class,
       AlertsEmailTemplate.class,
 
-      IssueFilterWsModule.class,
+    IssueFilterWsModule.class,
 
-      // action plan
+    // action plan
       ActionPlanWs.class,
       ActionPlanService.class,
 
-      // issues actions
+    // issues actions
       AssignAction.class,
       PlanAction.class,
       SetSeverityAction.class,
@@ -628,7 +626,7 @@ public class PlatformLevel4 extends PlatformLevel {
       AddTagsAction.class,
       RemoveTagsAction.class,
 
-      // technical debt
+    // technical debt
       DebtModelService.class,
       DebtModelOperations.class,
       DebtModelLookup.class,
@@ -638,7 +636,7 @@ public class PlatformLevel4 extends PlatformLevel {
       DebtRulesXMLImporter.class,
       DebtCharacteristicsXMLImporter.class,
 
-      // source
+    // source
       HtmlSourceDecorator.class,
       SourceService.class,
       SourcesWs.class,
@@ -652,23 +650,23 @@ public class PlatformLevel4 extends PlatformLevel {
       SourceLineIndex.class,
       SourceLineIndexer.class,
 
-      // Duplications
+    // Duplications
       DuplicationsParser.class,
       DuplicationsWs.class,
       DuplicationsJsonWriter.class,
       org.sonar.server.duplication.ws.ShowAction.class,
 
-      // text
+    // text
       MacroInterpreter.class,
       RubyTextService.class,
 
-      // Notifications
+    // Notifications
       EmailSettings.class,
       NotificationService.class,
       NotificationCenter.class,
       DefaultNotificationManager.class,
 
-      // Tests
+    // Tests
       CoverageService.class,
       TestsWs.class,
       CoveredFilesAction.class,
@@ -677,12 +675,12 @@ public class PlatformLevel4 extends PlatformLevel {
       TestIndex.class,
       TestIndexer.class,
 
-      // Properties
+    // Properties
       PropertiesWs.class,
 
-      TypeValidationModule.class,
+    TypeValidationModule.class,
 
-      // System
+    // System
       RestartAction.class,
       InfoAction.class,
       UpgradesAction.class,
@@ -697,7 +695,7 @@ public class PlatformLevel4 extends PlatformLevel {
       MigrateDbAction.class,
       DbMigrationStatusAction.class,
 
-      // Plugins WS
+    // Plugins WS
       PluginWSCommons.class,
       PluginUpdateAggregator.class,
       InstalledAction.class,
@@ -710,7 +708,7 @@ public class PlatformLevel4 extends PlatformLevel {
       CancelAllAction.class,
       PluginsWs.class,
 
-      // Compute engine
+    // Compute engine
       CEQueueStatusImpl.class,
       ComputeEngineQueueMonitor.class,
       ReportQueue.class,
@@ -724,7 +722,7 @@ public class PlatformLevel4 extends PlatformLevel {
       ProjectSettingsFactory.class,
       IndexPurgeListener.class,
 
-      // UI
+    // UI
       GlobalNavigationAction.class,
       SettingsNavigationAction.class,
       ComponentNavigationAction.class,
