+++ server/sonar-server/src/test/java/org/sonar/server/issue/InternalRubyIssueServiceTest.java
@@ -37,24 +37,17 @@ import org.sonar.api.web.UserRole;
 import org.sonar.core.issue.DefaultActionPlan;
 import org.sonar.core.issue.DefaultIssue;
 import org.sonar.core.issue.FieldDiffs;
-import org.sonar.db.DbClient;
-import org.sonar.db.DbSession;
 import org.sonar.db.component.ResourceDao;
 import org.sonar.db.component.ResourceDto;
 import org.sonar.db.component.ResourceQuery;
 import org.sonar.db.issue.IssueFilterDto;
-import org.sonar.server.component.ws.ComponentJsonWriter;
 import org.sonar.server.es.SearchOptions;
 import org.sonar.server.exceptions.BadRequestException;
 import org.sonar.server.exceptions.Message;
 import org.sonar.server.issue.actionplan.ActionPlanService;
 import org.sonar.server.issue.filter.IssueFilterService;
-import org.sonar.server.issue.ws.IssueComponentHelper;
-import org.sonar.server.issue.ws.IssueJsonWriter;
 import org.sonar.server.tester.UserSessionRule;
 import org.sonar.server.user.ThreadLocalUserSession;
-import org.sonar.server.user.index.UserIndex;
-import org.sonar.server.user.ws.UserJsonWriter;
 
 import static com.google.common.collect.Lists.newArrayList;
 import static com.google.common.collect.Maps.newHashMap;
@@ -91,20 +84,6 @@ public class InternalRubyIssueServiceTest {
 
   InternalRubyIssueService service;
 
-  IssueJsonWriter issueWriter;
-
-  IssueComponentHelper issueComponentHelper;
-
-  UserIndex userIndex;
-
-  DbClient dbClient;
-
-  DbSession dbSession;
-
-  UserJsonWriter userWriter;
-
-  ComponentJsonWriter componentWriter;
-
   @Before
   public void setUp() {
     issueService = mock(IssueService.class);
@@ -115,19 +94,12 @@ public class InternalRubyIssueServiceTest {
     resourceDao = mock(ResourceDao.class);
     issueFilterService = mock(IssueFilterService.class);
     issueBulkChangeService = mock(IssueBulkChangeService.class);
-    issueWriter = mock(IssueJsonWriter.class);
-    issueComponentHelper = mock(IssueComponentHelper.class);
-    userIndex = mock(UserIndex.class);
-    dbClient = mock(DbClient.class);
-    dbSession = mock(DbSession.class);
-    userWriter = mock(UserJsonWriter.class);
-    userWriter = mock(UserJsonWriter.class);
 
     ResourceDto project = new ResourceDto().setKey("org.sonar.Sample");
     when(resourceDao.selectResource(any(ResourceQuery.class))).thenReturn(project);
 
     service = new InternalRubyIssueService(issueService, issueQueryService, commentService, changelogService, actionPlanService, resourceDao,
-      issueFilterService, issueBulkChangeService, issueWriter, issueComponentHelper, componentWriter, userIndex, dbClient, userSessionRule, userWriter);
+      issueFilterService, issueBulkChangeService, userSessionRule);
   }
 
   @Test
