+++ server/sonar-server/src/test/java/org/sonar/server/issue/ws/IssueActionsWriterTest.java
@@ -33,7 +33,6 @@ import org.sonar.api.utils.text.JsonWriter;
 import org.sonar.api.web.UserRole;
 import org.sonar.core.issue.DefaultIssue;
 import org.sonar.core.issue.workflow.Transition;
-import org.sonar.server.issue.ActionService;
 import org.sonar.server.issue.IssueService;
 import org.sonar.server.tester.UserSessionRule;
 import org.sonar.test.JsonAssert;
@@ -50,14 +49,11 @@ public class IssueActionsWriterTest {
   @Mock
   IssueService issueService;
 
-  @Mock
-  ActionService actionService;
-
   IssueActionsWriter writer;
 
   @Before
   public void setUp() {
-    writer = new IssueActionsWriter(issueService, actionService, userSessionRule);
+    writer = new IssueActionsWriter(issueService, userSessionRule);
   }
 
   @Test
